Rework payroll work caps: monthly overtime (120h), Friday (16h x Fridays), night shift (8h x days), mission days <= standard days, holidays input (cap 24h each), remove absence, integer worked days

+++ src/Core/Contracts/PayrollRecords/PayrollRecordResult.cs
@@ -9,7 +9,7 @@ public record PayrollRecordResult(
     string DepartmentName,
     DateOnly PeriodStart,
     DateOnly PeriodEnd,
-    decimal WorkedDaysCount,
+    int WorkedDaysCount,
     decimal OvertimeHours,
     decimal GrossAmount,
     decimal TotalDeductionsAmount,
